🚚 MainContract 패키지를 app 모듈로
