+++ app/src/main/java/edu/nextstep/camp/calculator/MainContract.kt
@@ -1,0 +1,18 @@
+package edu.nextstep.camp.calculator
+
+import edu.nextstep.camp.calculator.domain.Expression
+import edu.nextstep.camp.calculator.domain.Operator
+
+interface MainContract {
+    interface View {
+        fun showExpression(expression: Expression)
+        fun showCalculationFailMessage()
+    }
+
+    interface Presenter {
+        fun addToExpression(operand: Int)
+        fun addToExpression(operator: Operator)
+        fun removeLast()
+        fun calculate()
+    }
+}
